Updated to version 0.2.1
Updated the Backend and the DashboardView and ItemView. This update adds a correct item managment and also allow for in grid edit while being persistent on the data. Furhtermore Dashboard now utilizes ApexCharts to show the donut chart of the individual categoes items. Smaller changes and bugfixed and been made. Furthermore we have gone from MySQL to PostGRES database.